requirements.txt
requests
beautifulsoup4
selenium
pymupdf
python-docx
openpyxl
nltk
spacy
langchain
faiss-cpu
sentence-transformers
streamlit
networkx
scikit-learn
neo4j 